src/requirements.txt
fastapi
uvicorn
httpx
google-generativeai
python-dotenv
tenacity
structlog
